fix getting username on windows
