--- user/user.go
@@ -1,44 +1,0 @@
-//go:build darwin
-
-package user
-
-import (
-	"os"
-	"os/user"
-	"strconv"
-)
-
-type User struct {
-	Uid      int
-	Gid      int
-	Username string
-	SudoRoot bool
-}
-
-func Current() User {
-	username := ""
-	sudoed := false
-	uid := os.Getuid()
-	gid := os.Getgid()
-	if uid == 0 {
-		// after a sudo, macOs returns the root user on both os.Getuid() and os.Geteuid()
-		// to detect the logged on user after a sudo, we need to use the environment variable
-		if userid, sudo := os.LookupEnv("SUDO_UID"); sudo {
-			if temp, err := strconv.Atoi(userid); err == nil {
-				uid = temp
-				sudoed = true
-			}
-		}
-	}
-	current, err := user.LookupId(strconv.Itoa(uid))
-	if err == nil {
-		gid, _ = strconv.Atoi(current.Gid)
-		username = current.Username
-	}
-	return User{
-		Uid:      uid,
-		Gid:      gid,
-		Username: username,
-		SudoRoot: sudoed,
-	}
-}
